requirements.txt
aiofiles==23.1.0
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.7.2
blinker==1.6.2
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.5
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.3
dnspython==2.4.0
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
Flask==2.3.2
greenlet==2.0.2
gyp==0.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httptools==0.6.0
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.9.1
protobuf==3.12.4
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyasn1==0.5.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.0.3
pydantic_core==2.3.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.4.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.23.0
uvloop==0.17.0
virtualenv==20.24.0
wadllib==1.3.6
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
